% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls.fit.R
\name{xls.fit}
\alias{xls.fit}
\title{Fit an eXtreme Least Squares Model}
\usage{
xls.fit(formula, data, error_weights = NULL, error_ahead_level = 4)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{A "data.frame" object containing the variables in the model.}

\item{error_weights}{A numeric vector including error weights by order. If NULL, it is created automatically by error_ahead_level amount, decreasing at equal intervals.}

\item{error_ahead_level}{An integer which represents how many steps further the parameters will be optimized for each data point.}
}
\description{
Almost the same interface as \code{base::lm}.
Just includes two parameters more, error_weights and error_ahead_level
}
