% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls.objfun.R
\name{xls.objfun}
\alias{xls.objfun}
\title{Preparing eXtreme Least Squares Nonlinear Objective Function}
\usage{
xls.objfun(data, error_column_name, error_weights, error_ahead_level)
}
\arguments{
\item{data}{A data.frame object which is returned by \code{xls.prep}. Tip: \code{xls.prep}'s \code{.$data} sub object returns the data.frame}

\item{error_column_name}{Symbolic error column's name. By default, it is named "error_symbolic" by \code{xls.prep()}}

\item{error_weights}{A numeric vector including error weights by order.}

\item{error_ahead_level}{An integer which represents how many steps further the parameters will be optimized for each data point.}
}
\value{
A \code{function} object.
}
\description{
Automatically used in \code{xls.fit()}
No need to use if the objective function is not specifically desired to be achieved.
}
