% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls.prep.R
\name{xls.prep}
\alias{xls.prep}
\title{Preparing eXtreme Least Squares Data}
\usage{
xls.prep(formula, data, dependent_var)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{A data.frame object.}

\item{dependent.var}{A character which is the same as left hand side variable in specified formula.}
}
\description{
Automatically used in xls.fit()
No need to use if the raw data is not specifically desired to be achieved.
}
