% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls.prep.R
\name{xls.prep}
\alias{xls.prep}
\title{Preparing eXtreme Least Squares Data}
\usage{
xls.prep(formula, data, dependent_var)
}
\arguments{
\item{formula}{An object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{A data.frame object.}

\item{dependent_var}{A character which is the same as left hand side variable in specified formula.}
}
\value{
A \code{list} object which contains a \code{data.frame} object to be modeled and character vector of independent variables.
}
\description{
Automatically used in \code{xls.fit()}
No need to use if the raw data is not specifically desired to be achieved.
}
